+++ NINA/Utility/Mediator/Interfaces/ICameraMediator.cs
@@ -23,6 +23,7 @@
 
 using NINA.Model;
 using NINA.Model.MyCamera;
+using NINA.ViewModel.Equipment.Camera;
 using NINA.ViewModel.Interfaces;
 using System;
 using System.Collections.Async;