+++ NINA/Utility/Mediator/Interfaces/IFilterWheelMediator.cs
@@ -23,6 +23,7 @@
 
 using NINA.Model;
 using NINA.Model.MyFilterWheel;
+using NINA.ViewModel.Equipment.FilterWheel;
 using NINA.ViewModel.Interfaces;
 using System;
 using System.Collections.Generic;