+++ NINA/ViewModel/Equipment/Guider/IGuiderChooserVM.cs
@@ -1,0 +1,10 @@
+﻿using NINA.Model.MyGuider;
+using NINA.Utility;
+
+namespace NINA.ViewModel.Equipment.Guider {
+
+    public interface IGuiderChooserVM {
+        AsyncObservableCollection<IGuider> Guiders { get; set; }
+        IGuider SelectedGuider { get; set; }
+    }
+}